Add Screenshots to README file

+++ BlogApp/Infrastructures/SeedDbContext.cs
@@ -17,12 +17,19 @@ namespace BlogApp.Infrastructures
 
             ILogger<DataBaseSeedler> logger = app.Services.GetRequiredService<ILogger<DataBaseSeedler>>();
 
-            dataDb.Database.EnsureDeleted();
+            if (app.Environment.IsDevelopment())
+            {
+                dataDb.Database.EnsureDeleted();
+                identityDb.Database.EnsureDeleted();                
+            } 
+            
             dataDb.Database.Migrate();
-            identityDb.Database.EnsureDeleted();
             identityDb.Database.Migrate();
 
-            await userProfilePictureService.RemoveAllProfilePicturesAsync();
+            if (app.Environment.IsDevelopment())
+            {
+                await userProfilePictureService.RemoveAllProfilePicturesAsync();
+            }
 
             if (!roleManager.Roles.Any())
             {
@@ -61,9 +68,9 @@ namespace BlogApp.Infrastructures
                 Blog blog4 = new Blog() { Title = "Programmer", BlogContent = "I want to be a great Backend programmer.", AppUserId = 3 };
                 Blog blog5 = new Blog() { Title = "Life", BlogContent = "I want to buy a BIG house.", AppUserId = 3 };
                 Blog blog6 = new Blog() { Title = "Life", BlogContent = "I want to buy a nice car (or nice super car).", AppUserId = 3 };
-                Blog blog7 = new Blog() { Title = "Life", BlogContent = "I want to make my parrents happy.", AppUserId = 4 };
-                Blog blog8 = new Blog() { Title = "Life", BlogContent = "I want to find nice GF and even possible marry her.", AppUserId = 4 };
-                Blog blog9 = new Blog() { Title = "Site", BlogContent = "I have built this site ;)", AppUserId = 1, IsConfirmed = true, IsPublic = true };
+                Blog blog7 = new Blog() { Title = "Family", BlogContent = "I want to make my parrents happy.", AppUserId = 1 };
+                Blog blog8 = new Blog() { Title = "Life", BlogContent = "I want to find nice GF and even possible marry her.", AppUserId = 2 };
+                Blog blog9 = new Blog() { Title = "My Website", BlogContent = "I have built this site ;)", AppUserId = 1, IsConfirmed = true, IsPublic = true };
 
                 blogs = [blog1, blog2, blog3, blog4, blog5, blog6, blog7, blog8, blog9];
 
@@ -71,12 +78,12 @@ namespace BlogApp.Infrastructures
                 dataDb.SaveChanges();
 
                 // create comments for some Blogs
-                Comment comment1 = new Comment() { StringContent = "God I love this Blog 1 from admin", AppUserId = 2, BlogId = 1 };
-                Comment comment2 = new Comment() { StringContent = "God I love this blog 1 from bob", AppUserId = 3, BlogId = 1 };
-                Comment comment3 = new Comment() { StringContent = "God I love this blog 1 from manager", AppUserId = 1, BlogId = 1 };
-                Comment comment4 = new Comment() { StringContent = "God I love this blog 4 from bob", AppUserId = 3, BlogId = 4 };
-                Comment comment5 = new Comment() { StringContent = "God I love this blog 1 from bob", AppUserId = 3, BlogId = 1 };
-                Comment comment6 = new Comment() { StringContent = "God I love this blog 2 from bob", AppUserId = 3, BlogId = 2 };
+                Comment comment1 = new Comment() { StringContent = "God, I love this Blog 1 from admin", AppUserId = 2, BlogId = 1 };
+                Comment comment2 = new Comment() { StringContent = "God, I love this blog 1 from bob", AppUserId = 3, BlogId = 1 };
+                Comment comment3 = new Comment() { StringContent = "God, I love this blog 1 from manager", AppUserId = 1, BlogId = 1 };
+                Comment comment4 = new Comment() { StringContent = "God, I love this blog 4 from bob", AppUserId = 3, BlogId = 4 };
+                Comment comment5 = new Comment() { StringContent = "God, I love this blog 1 from bob", AppUserId = 3, BlogId = 1 };
+                Comment comment6 = new Comment() { StringContent = "God, I love this blog 2 from bob", AppUserId = 3, BlogId = 2 };
 
                 Comment[] comments = [comment1, comment2, comment3, comment4, comment5, comment6];
